+++ arachne/scrapy_utils.py
@@ -37,28 +37,38 @@ def start_logger(debug):
         logger = ScrapyFileLogObserver(logfile_, logging.INFO)
         tlog.addObserver(logger.emit)
 
-def get_spider_settings(flask_app_config):
+def get_spider_settings(flask_app_config, spider_scrapy_settings):
     """
-    For the given spider_pipelines(dict) create a scrapy Settings object with
-    the common settings for each spider/crawler.
+    For the given settings for each spider create a scrapy Settings object with
+    the common settings spider & its personal settings. Individual spider scrapy
+    settings takes priority over global scrapy settings
 
     Returns:
         Scrapy settings class instance
-    """
-    settings = Settings()
 
-    # set the pipelines for exporting to JSON and CSV
+    .. vesion 0.3.0:
+       Allow settings for individual spiders and global settings
+    """
+    all_settings = {}
     pipelines = {}
     if flask_app_config['EXPORT_JSON']:
         pipelines['arachne.pipelines.ExportJSON'] = 100
     if flask_app_config['EXPORT_CSV']:
         pipelines['arachne.pipelines.ExportCSV'] = 200
-    settings.set('ITEM_PIPELINES', pipelines)
+
+    # allow user to pass settings for each spider if needed
+    all_settings['ITEM_PIPELINES'] = pipelines
+
+    # update spider global settings and personal settings
+    all_settings.update(flask_app_config['SCRAPY_SETTINGS'])
+    all_settings.update(spider_scrapy_settings)
+    settings = Settings(all_settings)
+
     return settings
 
-def start_crawler(spider_loc, flask_app_config):
+def start_crawler(spider_loc, flask_app_config, spider_scrapy_settings):
     start_logger(flask_app_config['DEBUG'])
     spider = load_object(spider_loc)
-    settings = get_spider_settings(flask_app_config)
+    settings = get_spider_settings(flask_app_config, spider_scrapy_settings)
     crawler = create_crawler_object(spider(), settings)
     crawler.start()