+++ CHANGES.md
@@ -4,6 +4,11 @@ Changelog
 Here you will find the full list of changes between each Arachne release
 
 
+Version 0.2.0 (Nov 23rd 2015)
+-----------------------------
+- Add individual spider settings to the `scrapy_settings` variable 
+- Add global spider settings to the `SCRAPY_SETTINGS` variable 
+
 Version 0.2.0 (Nov 15th 2015)
 -----------------------------